add review

+++ lib/models/homemodel.dart
@@ -48,7 +48,7 @@ class RecDest {
   int id;
   String name;
   int categoryId;
-  int rating;
+  double rating;
   String description;
   String address;
   String openDay;
@@ -61,7 +61,7 @@ class RecDest {
         id: json["id"],
         name: json["name"],
         categoryId: json["categoryId"],
-        rating: json["rating"],
+        rating: json["rating"].toDouble(),
         description: json["description"],
         address: json["address"],
         openDay: json["open_day"],
