+++ lib/models/wishlist_destination_model.dart
@@ -30,7 +30,7 @@ class WishlistDestination {
   int id;
   String name;
   int categoryId;
-  int rating;
+  double rating;
   String description;
   String address;
   String openDay;
@@ -45,7 +45,7 @@ class WishlistDestination {
         id: json["id"],
         name: json["name"],
         categoryId: json["categoryId"],
-        rating: json["rating"],
+        rating: json["rating"].toDouble(),
         description: json["description"],
         address: json["address"],
         openDay: json["open_day"],
